cgroup/v2 - Add no_file_cache to subtract fs-backed reclaimable memory
The active_file and inactive_file are reclaimable filesystem-backed memory,
and are included in memory.current and memory.peak.

The effect of counting these two is that a job which is simply creating
files in disk might get its memory reporting increased. That is misleading
as the cache pages might be evicted at any point, and it is a kernel
decision whether to use or not this cache. The user has no control over
it and the memory will be normally available after evicting it.

Add a no_file_cache option to JobAcctGatherParam to make it possible
to opt-out for including these in the accounted memory RSS and MaxRSS.

This will cause that memory.peak interface is no longer used so we are
back to a polling mechanism for MaxRSS. This might return old issues in
cgroup/v2 where short memory spikes were not detected between
jobacctgather intervals.

Ticket: 23646
Cherry-picked: 213a0e4b10