stepmgr - handle when step requested threads_per_core != node ThreadsPerCore
When calculating the usable_cpu_cnt of a node in _pick_step_nodes we
were not account for steps that request a non default threads per core,
such as when using --hint=nomultithread. This leads to "lost" steps and
a pause of SlurmctldTimeout from when the step was submitted before the
srun command times out and attempts to resubmits the step.

Here we are adjusting the usable_cpu_cnt to take this into account.

Changelog: stepmgr - handle when a steps requested ThreadsPerCore
 does not equal a nodes configured ThreadsPerCore
Ticket: 20799