Wait for systemd notify to be processed before exiting on reconfig
This aims to prevent a race that can happen when systemd attempts to
process a notify after the process ended. The barrier will keep the
process alive until the notification is processed, which will ensure
systemd is aware of the new PID for the daemon after reconfiguration.

Ticket: 23474
Changelog: Fix systemd potentially losing track of the daemon PIDs.