data_parser/v0.0.44 - In DUMP_FUNC(HOSTLIST_STRING) always return list
If not running in complex mode all DUMP_FUNC functions have to set the data
type of dst to the expected value even if it is empty. Before
DUMP_FUNC(HOSTLIST_STRING) was leaving the dst as data DATA_TYPE_NULL even
when not when running in complex mode. It should set it to DATA_TYPE_LIST
in that case.

Changelog: data_parser/v0.0.44 - Avoid hitting an xassert by return a
 DATA_TYPE_LIST on error when dumping a HOSTLIST_STRING without complex
 mode instead of returning DATA_TYPE_NULL.
Issue: 50547