Add ability to create dynamic nodes with gres.conf configuration
node_record_t on the controller is fully setup when the node registers
with the controller.

Cloud nodes need to load in the gres.conf on the controller because
nodes can be allocated before they register with the node's gres
information.

External nodes never have a slurmd attached to them, so the controller
will never get a gres response from them. In order to provide scheduling
of indexed gres, the gres_slurmd_conf_t information needs to be
provided.