accounting_storage/mysql - Fix user name case with PreserveCaseUser
When PreserveCaseUser had not been set and an existing user over 24hrs
old (and/or one that had run jobs) was deleted and PreserveCaseUser
was set and the user was added with a different case then the user
table entry for the user was not updated properly to match the assoc
table one. This caused "sacctmgr -s show user ..." to incorrectly list
details for the user.

Similarly, when PreserveCaseUser had been set and an existing user was
deleted and PreserveCaseUser was cleared and the user was added with a
different case then the user table entry for the user was not properly
updated. This patch fixes these two cases.

Ticket: 23675