slurmd - change error to debug the energy rpc burst message
When an energy rpc burst occurs, print as an error the first "Too many
pending requests" message and then switch to debug logging while it
continues to happen to avoid flooding the log. When all the pending
requests have been processed, switch to error again for the first next
occurrence.

Ticket: 23600