slurmctld - cancel interactive jobs that never ran the prolog
Continuation of previous commit.

If prolog RPC never reached its destination, but we restarted the slurmctld,
we lose track of the RPC and the job is not cancelled.

This commit inserts a logic in the purge operation that is able to cancel
those jobs once slurmd is restarted, or every time
MESSAGE_NODE_REGISTRATION_STATUS is sent to the controller.

We need to wait for the EpilogTimeout here to ensure we cancel only the right
jobs.

Ticket: 23720
Changelog: slurmctld - Cancel interactive jobs that never ran the prolog in
 the purge jobs logic.
Cherry-picked: a39d8327b8