Testsuite - Fix test7.10 (1/2) for TLS adding setresgid to preload file
test7.10 tests if we can fake being root by preloading a .so to sbatch
that redefines "get uid" type functions to always return root id (0).

In a TLS environment, user commands generate a ephemeral TLS cert/key
pair in case they need to open a listening socket for others to connect
to (In TLS environment, clients always need a TLS cert to trust when
connecting to a "server", and in this case, the user command could be
acting as a "server").

In order to generate these ephemeral certificates, user commands may
use the certgen/script plugin which basically executes small shell
scripts containing openssl commands to generate a cert/key pair. The
cergen/script plugin uses the run_command() system.

run_command() calls setresgid() and setresuid() in order to sync the
real uid with the effective uid by calling geteuid()/getegid(). See
81e2648ac4d for the history on this.

Since test7.10 is hacking geteuid()/getegid() in sbatch to pretend to be
root, setresgid()/setresuid() break because they will error with EPERM
if they try to set uid to root without proper permissions. In this case,
we can just hack setresgid()/setresuid() to do nothing to avoid this
error. No-op setresgid()/setresuid() does not break this test, as the
gid/uid are arbitrarily set via the hacked .so, and it does not rely on
setresgid()/setresuid() to hack the gid/uid.

This is a partial fix (1/2), next commit should complete the fix.