accounting_storage/mysql - Preserve case on user remove/modify
When PreserveCaseUser has been set and a differing case version of
an existing user is specified when removing or modifying a user, the
specified case of the user is displayed rather than the original case of
the existing user. Fix that by displaying the user with the original case.

Changelog: Fix user name case not being preserved in user and assoc
 tables when toggling PreserveCaseUser, or when modifying/removing a
 user with a different case than stored.
Ticket: 23675