slurmctld - Don't allow manual node updates to IDLE if in REBOOT_ISSUED
Prevent DOWN+REBOOT_ISSUED nodes from being manually set to
IDLE+REBOOT_ISSUED. The REBOOT_ISSUED state only gets automatically cleared
if the node is in the DOWN state. Also not allowing nodes to be idled
prevents jobs being allocated on a node that is about to reboot.

Changelog: Manually updating a node's state to IDLE with 'scontrol update
 nodename' is no longer allowed if the node is in state REBOOT_ISSUED. This
 is to prevent the REBBOOT_ISSUED state from not automatically clearing
 once the reboot finishes.
Ticket: 23823