Add a global pointer to the stepd_step_rec_t
This structure is foundational to everything slurmstepd does.
Make it globally accessible so a lot of calling paths can be simplified.
This does preclude the ability to manage multiple steps within a single
slurmstepd (which I'm assuming was the concept behind this having been
explicitly managed), but there are many other things that block that
design from being fulfilled at this point.