metrics/openmetrics - Implement plugin logic
This plugin allows a user to parse metrics structs into metric sets, and to
dump metric sets into a string buffer in OpenMetrics output format
specification.

It uses two hash tables stored in the metrics set to sort metrics. One table
stores all the metrics based on an identifier generated by the metric name,
and the set of key-values. Another table stores a set of lists ordered by
metric name where every list contains the metrics with the same name.

The API is adapted to Slurm needs and does not fully implement the
OpenMetrics model and specification.

Issue: 50538