Print error message if input node state is invalid
_node_state_id outputs -1 on invalid node state input. When previously
stored in id->state, the output of _node_state_id goes to an unsigned
int, so when it returns a -1 it overflows. This prevents the error
message from printing, and sinfo --state instead reports all nodes as
having state n/a.

Changelog: sinfo --state/-t flag now outputs a more helpful error
 message when passed an invalid node state.
Ticket: 23752