Testsuite - Fix test7.10 (2/2) for TLS changing ownership of ca_cert.pem
In test7.10, 'unshare -ru' is used to attempt a fake root attack. When
this is done, all files owned by the user appear as owned by 'root' and
all other files will appear as owned by 'nobody'. The tls/s2n plugin
verifies the ownership of the CA certificate to be either root or
SlurmUser. In this case with 'unshare -ru' however, the file will appear
to be owned by 'nobody', which causes Slurm commands to fail.

A malicious user could easily circumvent this issue by creating their
own ca_cert.pem file and pointing Slurm to that.

For the purposes of this test, we can simulate this ca_cert.pem file
ownership change by actually changing the ca_cert.pem ownership
(something a malicious user might not be able to do). This allows this
test to still run.

This complete the fix (2/2) started in the previous commit.