Typecast to avoid coverity issues with sign extension
Continuation of 88ac8ab7a and eeaed893f3 which neither of them
fixed the issues.

Use uint64_t for i instead of the same type as node_ptr->tpc (uint16_t) because
multiplying two uint16_t values is promoted to a signed int.

This implicit promotion can lead to unintended sign extension when the result is
later converted to int64_t in bit_nset() arguments.

Declaring i as uint64_t ensures the multiplication is done in an unsigned 64-bit
context, preventing sign extension and Coverity warnings.

We also cast to bitstroff_t to avoid Coverity complaints, as it seems there can
be a difference between an implicit cast and explicit cast in the eyes of
Coverity. While explicit is fine, implicit could be unintentional.

CID: 521407 521389
Issue: 50456