Add REQUEST_GET_BPF_TOKEN to get a BPF token
This RPC is meant to only be served from a extern step which is not on a
user namespace, the ones in a user namespace will use it to contact the
extern slurmstepd, which is also outside the user namespace and get a
BPF token from it. The token gets generated the first time it is
requested, the next times, as it is stored in the cgroup plugin it is
directly delivered.

Issue: 50542