sdiag - Fix error when printing rpc stats with too long hostlists
If the hostlist string is too long, then sdiag with the json output format
will fail with an error. The hostlist would be truncated to 80 char and
parsing can fail if the string terminates with an unmatched bracket.

Fix it by allocating the hostlist strings.

Ticket: 23937
Cherry-picked: 75c20e914c