Remove _acct_restore_active_jobs()
This is vestigal. The only way to trigger this would be to fail to load
the job_state file, but with job_list already populated with jobs.

Note that load_all_job_state() will fatal before load_job_ret could be
non-zero unless ignore_state_errors is set. The only situation in which
read_slurm_conf() is ever called repeatedly in slurmctld at this point
is on an HA hand-off cycle. So you would need to have a failover happen
while running with -i, and for the state file to be unavailable or corrupt,
and for the job records in memory to still be valid.

In this situation, the in-memory job queue is likely incorrect. And running
an HA slurmctld with -i is already ill-advised.