Testsuite - Fix test7.10 for s2n with mount namespace for ca_cert.pem
Previously, chown was used to change ownership of ca_cert.pem which
would bypass the check that srun did to see if ca_cert.pem was owned by
SlurmUser/root.

Prior to 3553ecf43a, this worked fine as only client commands needed to
load ca_cert.pem for the test. After this commit though, slurmstepd also
needs to load ca_cert.pem, in which case it would fail since it was now
owned by the testsuite_user and not SlurmUser/root.

This commit creates a new mount just for srun so that it can pretend
that it owns ca_cert.pem, which will appear as root in the new user
namespace and avoid any checks on its ownership.