data_parser/v0.0.4[2,3,4]/parsers.c - mark SORT as a removed flag
The SORT flag was removed, but that broke the openapi specs
that have been snapshotted and expect the flag to be there. This adds
SORT back in and marks it as "removed" so that the tests pass, but the
functionality is still gone.

Regression from 700de4a2eb.

Ticket: 23807
Cherry-picked: a5a96060b5