Prevent memory corruption for large node counts and TreeWidth
It is possible to overflow the level_capacity and branch_capacity values in
_set_span() since they are integers and set to tree_width^(x iterations).
If TreeWidth >= 46341 and the cluster node count is >= (TreeWidth + 1) this
can reproduce. Overflowing these values can result in indexing out of range
of the 'span' array causing memory corruption.

Changelog: slurmctld - Prevent memory corruption when fanning out messages
 to the slurmds if TreeWidth is more then or equal to 46341 and the number
 of nodes in the cluster is more then or equal to (TreeWidth + 1).
Ticket: 23827