Fix hanging when fetching slurm.conf
When relying on the fetch_config function to retrieve a slurm.conf,
if the child process responsible for it dies before writing anything
to the pipe connecting with its parent process, the parent process
will be stuck forever.

This commit closes the unused pipe ends on both processes to avoid
having a read getting blocked indefinitely.

Ticket: 24049
Changelog: Fix issue where a cli command or process could get stuck
 indefinitely when trying to retrieve a slurm.conf from slurmctld.