Testsuite - Fix test_116_26 avoiding/catching false/real failures
We cannot down the BatchHost, and we should assert that the job is still
running and not only "not completing" (it could already failed).
Also, we should wait until node is really set to down to assert.

This is actually a fix/continuation of commit a53ab0809b.

Ticket: 24107
Ticket: 24119