Testsuite - Improve save+setup+restore of SysConfDir using new etc.orig
Instead of trying to save+restore any individual change that we do in any
config file, we should better use the same approach that we follow for
StateSaveLocation or the DB: save current one, setup a fresh one, and
restore the saved one once the test is over.

Note that this only works for auto-config and that it needs the
improvements done in the ATF repo, who is responsible to create the base
SysConfDir (etc.orig).

The previous backup_config_file() and remove_config_file() has been
removed.