slurmctld - Add error handling for malformed REQUEST_CONFIG RPC
Currently, when using configless if a slurmd sends a malformed config
request, an error is returned but it is thrown away causing a segfault
later on when the message is checked. This commit ensures that the error
is properly handled by returning it instead of throwing it away.

Changelog: slurmctld - Fixed segfault when running configless and a
 malformed REQUEST_CONFIG RPC is received.
Ticket: 23995
Cherry-picked: 1b91fac9