Fix cron/requeued jobs being incorrectly reported as runaway
If the "sacctmgr show runaway" is ran at the same moment a job is
requeued, or a cronjob resubmitted, it may be incorrectly reported as
runaway.

This is because until the new {JobID,Submit_time}pair has made it to the
database, controller and DB records won't be in sync.

Ticket: 24130
Changelog: Fix cron/requeued jobs being incorrectly reported as runaway
Cherry-picked: 8ada774347