Docs - Reword purging recommendation in high throughput guide
This clarifies that purging (a) is strongly recommended due to a large
number of jobs, and (b) that the recommendation is due to the manageability
of large tables in general, not due to any direct impact to job throughput.
It also removes the example since the right configuration is highly
situational.

Cherry-picked: f3e69a2dec