Testsuite - Improve test1.69 avoiding false failures
Parallel jobs needed 50s to complete and had 1m of time limit.
That should be fine, but in very busy shared systems they may be
cancelled.
Better reduce the run time of the jobs to only 10s, and even give
them more time.

Also, ensure that jobs really run in parallel and not sequentially by
forcing to catch both initial lines before the timers are triggered.