Fix possible overflow on CpuSpecList
We were setting the max bytes to be equal to num of cpus, but the
string representing one cpu can be up to 4 chars if max cpus is 1024.

In the worst case if we add commas and a terminator it is 5 * max cpus,
but if we want to do it dynamic we need to calculate the number of possible
digits in the string, add the commas and terminator.

In the unlikely case that the CpuSpecList was larger than 1024, the string
could be truncated and the constraint not applied.

Ticket: 22707
Changelog: Fix not applying constraints if CpuSpecList string is larger
 than 1024 chars.