Add error handling for malformed RESPONSE_CONFIG RPC
Currently, when fetching config files, for example when using configless,
if a malformed response to a config request is sent from the slurmctld it
just fails to fetch configs with no error. This commit helps capture the
error and log that a malformed RPC has been received.

Changelog: Added error handling and logging when a malformed
 RESPONSE_CONFIG RPC is received.
Ticket: 23995