cgroup/v2 - Avoid calling step create multiple times
This caused a memleak when each plugin called step_create() and did not
free the previous cgroup. In v1 that made sense as every cgroup was
different but in v2 we have a unified architecture and the cgroup is
the same for all plugins.

Changelog: Fix slurmstepd memleak when initializing cgroup plugins.
Ticket: 24225