Stop jobs that exceed magnetic reservation end time from being attracted
A magnetic reservation will pick up long running jobs that cannot fit in
the time limit of the reservation. This commit checks if a job can fit in
a magnetic reservation's time limit. If not, an error is thrown and the
job will not be scheduled under that reservation.

Changelog: Prevent jobs that cannot fit in the reservation's time limit from
 being attracted to a magnetic reservation.
Ticket: 24140