Remove incorrect unneeded code for RestrictedCoresPerGPU
From the original implementation of RestrictedCoresPerGPU with shared
gres.

This code attempts to force allocation of shared gres to match restricted cores
on their sockets. However it fails in many cases where it should allow
allocation.

It is overly complicated and beyond the scope of what we are attempting
to do here anyway.

This is a partial revert of 89de0158769

Changelog: Fix issue with RestrictedCoresPerGPU with shared gres.
Ticket: 22391
Cherry-picked: ff61b26257