Send job_start message if job_complete message is sent first
With the introduction of the sluid, jobs now have a db_index when they are
created as opposed to having a db_index of 0. Since the db_index is now
set as_mysql_job_complete() will always skip over the logic calling
as_mysql_job_start(). This could cause jobs to be marked as runaway. This
commit checks if the job is in the database and if not, will call
as_mysql_job_start().

Ticket: 24077
Changelog: Jobs that complete quickly will not be marked as runaway.