Testsuite - Improve test1.91 and test1.92 avoiding false failures
Sometimes srun or salloc commands may print expected error: lines if
some process in the system (even not related to Slurm) sends unexpected
messages to it, and this is correct behavior.

Therefore, instead of checking for error lines in the output, we should
use the actual exit code of srun in the test.