Unblock SIGCHLD before executing each task
SIGCHLD is blocked by the slurmstepd so that all slurmstepd threads
inherit the mask and go through the central signal handler.

An ignored SIGCHLD may or may not get reset to default when execve() is
called, and on Linux it is not which meant that all tasks had SIGCHLD
blocked by default.

Continuation of commit d32201a4e8.

Changelog: Fix SIGCHLD not being sent to tasks.
Ticket: 24294