conmgr - Cleanup TLS connection before closing underlying fd
In order to successfully close a TLS connection, a shutdown needs to
occur which may involve reading/writing by both parties. This needs to
happen before the underlying connection is closed.

For most Slurm communications, connections are finished after an RPC
response is received, meaning there's no need to check for a successful
TLS shutdown.

Other communications that are persistent in nature, meaning that there
are valid cases for the connection to receive more data or for the peer
to close the connection (e.g. checking for 0 from recv()), require the
successful TLS shutdown in order to distinguish intentional connection
closures versus errors resulting in connection closure.

This commit ensures that when a connection is ready to be closed, that
if there is a TLS state on it, an attempt to close the TLS state will
happen if the underlying connection is still open. When the TLS state is
successfully closed, then the underlying connection will be closed.

Issue: 50453