Remove shutdown from conn_g_destroy and remove graceful_shutdown
conn_g_set_graceful_shutdown() was added in an attempt to make TLS
shutdown more convenient by always calling shutdown inside of
conn_g_destroy() if needed. However, sometimes it is desirable to allow
callers to shutdown the TLS connection before actually de-allocating the
connection. For this reason, conn_g_shutdown() was added.

Using conn_g_set_graceful_shutdown() to enable this shutdown was making it
more difficult to see when TLS shutdown actually occurred for which
connection. Requiring an explicit call to conn_g_shutdown() or
conn_blocking_g_shutdown() makes this easier to follow.

Currently, conmgr does not handle any TLS connections that need to do a
shutdown, but in the future, the non-blocking conn_g_shutdown() will be
required in order to maintain asynchronous behavior on TLS connections,
and removing the shutdown from tls_p_destroy_conn and with the addition
of tls_g_shutdown_conn, this is expected to be easier.

Issue: 50453