Reset garbage SLUID from older versioned slurmd registration messages
When slurmd is preparing a slurm_node_registration_status_msg_t, it
creates a list of all slurm_step_id_t's that it detects on the node.

For slurmd's that are 25.05 or older, they will send uninitialized
garbage data in the slurm_step_id_t as part of
slurm_node_registration_status_msg_t, and slurmctld will interpret this
as an invalid sluid and abort the job.

This commit ensures that the garbage sluid is set to 0, which other
logic in slurmctld is designed to recognize as an older protocol version
job.

Ticket: 24312
Changelog: Fix issue with jobs running on slurmd's with version 25.05.x
 or older getting aborted when slurmd re-registers with slurmctld.